refactor(charts): extract value-label methods into ChartValueLabelMixin
Move the value-label normalize/format/render cluster out of Chart into a
new ChartValueLabelMixin, added as the first base so the methods resolve
unchanged via MRO. Method bodies are byte-for-byte identical; rendered
output is unchanged.
